# openapi.yaml
openapi: 3.0.3
info:
  title: Saukele Backend API
  version: 2.0.0
  description: Production-oriented API for wedding profiles, gift funding, kinship-based gift access, email verification, contributions, currencies, and queue visibility.
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /gifts:
    get:
      tags: [Gifts]
      summary: List current couple gifts
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Paginated gifts
    post:
      tags: [Gifts]
      summary: Create gift for current couple
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GiftInput"
            example:
              name: "Refrigerator"
              description: "Samsung two-door refrigerator"
              targetAmount: 350000
              currency: "KZT"
              allowedTiers: ["parents", "siblings"]
      responses:
        "201":
          description: Gift created
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /family/gifts:
    get:
      tags: [Family]
      summary: List gifts visible through current user's kinship
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Paginated visible gifts
...
  /contributions:
    post:
      tags: [Contributions]
      summary: Contribute to a visible gift
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContributionInput"
            example:
              giftId: 1
              amount: 25000
              currency: "KZT"
              isAnonymous: false
      responses:
        "201":
          description: Contribution completed and email notification queued
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      in: query
      name: page
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
  responses:
    ValidationError:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Forbidden:
      description: Authenticated user is not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    GiftInput:
      type: object
      required: [name, targetAmount]
      properties:
        name:
          type: string
        description:
          type: string
        targetAmount:
          type: integer
          minimum: 1000
        currency:
          type: string
          enum: [KZT, EUR, USD]
        allowedTiers:
          type: array
          items:
            type: string
            enum: [ata_ana, zhien_zaran, kuda_zhekzhen]
        imageUrls:
          type: array
          items:
            type: string
    ContributionInput:
      type: object
      required: [giftId, amount]
      properties:
        giftId:
          type: integer
        amount:
          type: integer
          minimum: 1
        currency:
          type: string
          enum: [KZT, EUR, USD]
        isAnonymous:
          type: boolean
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time